pyproject.toml
[project]
name = "mcp-example"
version = "0.1.0"
description = "A LangChain MCP application for mathematical calculations"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.9"
dependencies = [
    "langchain-mcp-adapters>=0.0.2",
    "langgraph>=0.0.1",
    "langchain-openai>=0.0.1",
    "python-dotenv>=0.19.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build]
include = [
    "src/**/*.py",
    "tests/**/*.py",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 